requirements.txt
python-dotenv
streamlit
google-generativeai
SpeechRecognition
pyaudio
pipwin
gtTS    

